feat: useForm state includes missing choice error
Initial pass at a useForm state that contains error message. There are a
lot of other error states we could add, but for now it only includes an
error if a user types an amount into the text input field but doesn't
include a radio button selection.

fix: error using UseFormState in client component