feat: working popup on desktop
What made this PR difficult was that we need to formActions, one
for validating the form before opening the popup, and another for
submitting the form from the popup.

Note there is still a bug where this isn't working properly on mobile.

Also we don't redirect the user after confirmation.