fix: mobile order drawer not working
The issue was because we render both the mobile and desktop versions but
just hide one of them. This means there are two of the same forms with
the same input elements and ids. So when a user presses the radio
buttons in the mobile version, it affects the hidden desktop version of
the radio buttons.