feat: use submarket name over choice_market_id
What was intended to be a small change ended up being a relatively big
refactor.

Motivation: Before, we identified  each line in the chart using
choice_market_id since it was the main identifier in price_histories
table. We wanted to switch to using the actual name of the submarket.

Changes include pulling all the data in a single query (instead of using
two separate queries) and changing the client-side formatting functions
to convert it into a useful data structure.